+++ server.js
@@ -158,8 +158,8 @@ app.get("/api/usage", (req, res) => {
   });
 });
 
-const ver = "v0.6.3";
-const sub = "(2023-11-03)";
+const ver = "v0.6.4";
+const sub = "(2023-12-22)";
 
 import { marked } from "marked";
 import { gfmHeadingId } from "marked-gfm-heading-id";