+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This document will update alongside the changelogs that come with each new versi
     - [Known Bugs](#known-bugs-1)
   - [v0.6.2](#v062)
   - [v0.6.3](#v063)
+  - [v0.6.4](#v064)
 
 ## v0.3.7
 
@@ -160,3 +161,8 @@ This document will update alongside the changelogs that come with each new versi
 ## v0.6.3
 
 To keep consistency with git, I accidentally wrote 0.6.3 as the 0.6.2 changelog. Whoops.
+
+## v0.6.4
+
+- Added import and export buttons to the settings menu
+- WIP language support, will be coming in a future date